Controller: Employeer Controller refactored after auth

+++ resources/views/admin/modules/employee/listemployee.blade.php
@@ -5,11 +5,9 @@
     <div class="row">
         <div class="col-md-3"></div>
         <div class="col-md-6">
-
             <div class="row">
                 <div class="col-md-3"></div>
                 <div class="col-md-6">
-
                     <h3 style="text-align: center;">Our Registered Users</h3>
                 </div>
                 <div class="col-md-3">
@@ -17,7 +15,6 @@
                 </div>
             </div>
         </div>
-
     </div>
     <thead class="bg-light">
         <tr>
@@ -32,49 +29,40 @@
         <tr>
             <td>
                 <div class="d-flex align-items-center">
-                    <img src="https://mdbootstrap.com/img/new/avatars/8.jpg" alt="" style="width: 45px; height: 45px" class="rounded-circle" />
+                    <img src="{{ asset('images/employee/profile/' . $employee->employee_image) }}" alt="" style="width: 45px; height: 45px" class="rounded-circle" />
                     <div class="ms-3">
-                        <p class="fw-bold mb-1">{{$employee->employee_first_name}}</p>
-                        <p class="text-muted mb-0">{{$employee->employee_email}}</p>
+                        <p class="fw-bold mb-1">{{ $employee->user->name }}</p>
+                        <p class="text-muted mb-0">{{ $employee->user->email }}</p>
                     </div>
                 </div>
             </td>
             <td>
-                <p class="fw-normal mb-1">{{$employee->employee_address}}</p>
-                <p class="text-muted mb-0">{{$employee->employee_contact}}</p>
+                <p class="fw-normal mb-1">{{ $employee->employee_address }}</p>
+                <p class="text-muted mb-0">{{ $employee->user->contact }}</p>
             </td>
-            @if($employee->employee_status == 'Inactive')
-            <?php $color = 'danger'; ?>
-            @elseif($employee->employee_status == 'under verification')
-            <?php $color = 'danger'; ?>
-            @else
-            <?php $color = 'success'; ?>
-            @endif
+            @php
+            $color = match($employee->employee_status) {
+            'Inactive', 'under verification' => 'danger',
+            default => 'success'
+            };
+            @endphp
             <td>
-                <a href="/employee/verification/{{$employee->id}}" onclick="return confirmStatusChange('change status')">
-                    <span class="badge badge-{{$color}} rounded-pill d-inline">{{$employee->employee_status}}</span>
+                <a href="/employee/verification/{{ $employee->id }}" onclick="return confirmStatusChange('change status')">
+                    <span class="badge badge-{{ $color }} rounded-pill d-inline">{{ $employee->employee_status }}</span>
                 </a>
             </td>
-
-            <script>
-                function confirmStatusChange(a) {
-                    return confirm("Are you sure you want to" + " " + a + "?");
-                }
-            </script>
             <td>
-                <a href="/employee/edit/{{$employee->id}}">
-                    <button type="button" class="btn btn-primary btn-sm btn-rounded mx-2 px-2 ">
+                <a href="/employee/edit/{{ $employee->id }}">
+                    <button type="button" class="btn btn-primary btn-sm btn-rounded mx-2 px-2">
                         Edit
                     </button>
                 </a>
-                <a href="/employee/delete/{{$employee->id}}">
-
-                    <button type="button" class="btn btn-danger btn-sm btn-rounded mx-2 px-2" onclick="return confirmStatusChange('delete')">
+                <a href="/employee/delete/{{ $employee->id }}" onclick="return confirmStatusChange('delete')">
+                    <button type="button" class="btn btn-danger btn-sm btn-rounded mx-2 px-2">
                         Delete
                     </button>
                 </a>
-                <a href="/employee/profile/{{$employee->id}}">
-
+                <a href="/employee/profile/{{ $employee->user_id }}">
                     <button type="button" class="btn btn-warning btn-sm btn-rounded mx-2 px-2">
                         Details
                     </button>
@@ -84,4 +72,10 @@
         @endforeach
     </tbody>
 </table>
+
+<script>
+    function confirmStatusChange(action) {
+        return confirm(`Are you sure you want to ${action}?`);
+    }
+</script>
 @endsection
